Retry svn checkout stage 5 times
On a recent build of the AppImage, the pipeline failed due to svn
failing with E210004: Malformed network data.
This appears to be an intermittent issue, as subsequent builds succeeded
with no changes to the repository. Additionally, it is out of our
control, since it is dependent on the sourceforge repository and the
network co-operating.